refactor: move FieldEncryptor from SDK to agentos-service internal package
FieldEncryptor is an internal service concern — plugins receive
pre-decrypted credentials via CredentialProvider and never interact
with encryption directly. No reason for this interface to be in the SDK.
