+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/encryption/NoOpFieldEncryptor.kt
@@ -1,7 +1,5 @@
 package io.whozoss.agentos.encryption
 
-import io.whozoss.agentos.sdk.encryption.FieldEncryptor
-
 /**
  * No-op field encryptor for development use only.
  * Values are stored as-is without encryption.
