+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/encryption/SpringFieldEncryptor.kt
@@ -1,6 +1,5 @@
 package io.whozoss.agentos.encryption
 
-import io.whozoss.agentos.sdk.encryption.FieldEncryptor
 import org.springframework.security.crypto.encrypt.Encryptors
 import org.springframework.security.crypto.encrypt.TextEncryptor
